backend/requirements.txt
# Existing packages (keep as is)
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ============ ML PACKAGES (ADD THESE) ============

# Statistical Models
statsmodels==0.14.1          # SARIMA
prophet==1.1.5               # Facebook Prophet (might need extra steps)

# Traditional ML
xgboost==2.0.3              # XGBoost for ensemble
scikit-learn==1.4.0         # Preprocessing, metrics

# Data Processing
pandas==2.2.0               # DataFrames
numpy==1.26.3               # Numerical operations

# Utilities
joblib==1.3.2               # Model serialization
tqdm==4.66.1                # Progress bars